+++ north-sea/next/package.json
@@ -2,8 +2,11 @@
   "dependencies": {
     "@react-navigation/native": "^6.1.1",
     "@react-navigation/native-stack": "^6.9.6",
+    "@reduxjs/toolkit": "^1.9.1",
     "@types/react": "^18.0.26",
     "@types/react-native": "^0.70.8",
+    "@types/react-redux": "^7.1.24",
+    "@types/redux": "^3.6.0",
     "expo": "^47.0.0",
     "next": "^13.0.7",
     "next-fonts": "^1.5.1",
@@ -16,6 +19,8 @@
     "react-native-screens": "^3.18.2",
     "react-native-svg": "13.4.0",
     "react-native-web": "~0.18.7",
+    "react-redux": "^8.0.5",
+    "redux": "^4.2.0",
     "typescript": "^4.6.3"
   },
   "devDependencies": {
