+++ north-sea/next/src/store.ts
@@ -1,0 +1,8 @@
+  import { configureStore } from "@reduxjs/toolkit";
+import mainReducer from './reducer'
+
+export type root = ReturnType<typeof store.getState>
+
+const store = configureStore({ reducer: mainReducer });
+
+export default store
