--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "typescript": "^4.9.4"
-  }
-}
